[actor] Migrate SyncJobsActor to use ActixWrapper
