+++ integration-tests/src/tests/client/features/multinode_test_loop_example.rs
@@ -37,8 +37,8 @@ use near_client::client_actions::{
     SyncJobsSenderForClientMessage,
 };
 use near_client::sync::sync_actor::SyncActor;
-use near_client::sync_jobs_actions::{
-    ClientSenderForSyncJobsMessage, SyncJobsActions, SyncJobsSenderForSyncJobsMessage,
+use near_client::sync_jobs_actor::{
+    ClientSenderForSyncJobsMessage, SyncJobsActor, SyncJobsSenderForSyncJobsMessage,
 };
 use near_client::test_utils::test_loop::client_actions::{
     forward_client_messages_from_client_to_client_actions,
@@ -54,9 +54,9 @@ use near_client::test_utils::test_loop::sync_actor::{
     forward_sync_actor_messages_from_client, forward_sync_actor_messages_from_network,
     test_loop_sync_actor_maker, TestSyncActors,
 };
-use near_client::test_utils::test_loop::sync_jobs_actions::{
-    forward_sync_jobs_messages_from_client_to_sync_jobs_actions,
-    forward_sync_jobs_messages_from_sync_jobs_to_sync_jobs_actions,
+use near_client::test_utils::test_loop::sync_jobs_actor::{
+    forward_messages_from_client_to_sync_jobs_actor,
+    forward_messages_from_sync_jobs_to_sync_jobs_actor,
 };
 use near_client::test_utils::test_loop::{
     forward_messages_from_partial_witness_actor_to_client,
@@ -100,7 +100,7 @@ struct TestData {
     pub dummy: (),
     pub account: AccountId,
     pub client: ClientActions,
-    pub sync_jobs: SyncJobsActions,
+    pub sync_jobs: SyncJobsActor,
     pub shards_manager: ShardsManager,
     pub partial_witness: PartialWitnessActions,
     pub sync_actors: TestSyncActors,
@@ -289,7 +289,7 @@ fn test_client_with_multi_test_loop() {
         let store = opener.open().unwrap().get_hot_store();
         initialize_genesis_state(store.clone(), &genesis, None);
 
-        let sync_jobs_actions = SyncJobsActions::new(
+        let sync_jobs_actor = SyncJobsActor::new(
             builder
                 .sender()
                 .for_index(idx)
@@ -434,7 +434,7 @@ fn test_client_with_multi_test_loop() {
             dummy: (),
             account: accounts[idx].clone(),
             client: client_actions,
-            sync_jobs: sync_jobs_actions,
+            sync_jobs: sync_jobs_actor,
             shards_manager,
             partial_witness: partial_witness_actions,
             sync_actors,
@@ -476,14 +476,14 @@ fn test_client_with_multi_test_loop() {
 
         // Messages to the SyncJobs component.
         test.register_handler(
-            forward_sync_jobs_messages_from_client_to_sync_jobs_actions(
+            forward_messages_from_client_to_sync_jobs_actor(
                 test.sender().for_index(idx).into_future_spawner(),
             )
             .widen()
             .for_index(idx),
         );
         test.register_handler(
-            forward_sync_jobs_messages_from_sync_jobs_to_sync_jobs_actions(
+            forward_messages_from_sync_jobs_to_sync_jobs_actor(
                 test.sender().for_index(idx).into_future_spawner(),
             )
             .widen()
